requirements.txt
swarmauri[full]
PyPDF2
python-dotenv
gradio